Added Autonginx 0.0.1
It's preety much working. The only issue I've encountered that during creating of nginx user it just crashes, because nginx user is busy with process. Deleting that user and relaunching playbook fixes the issue.

Don't forget to change the key in authorized_keys file and your IP in hosts file